# server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Chronograf",
    "description": "API endpoints for Chronograf",
    "version": "1.8.4"
  },
  "schemes": ["http"],
  "basePath": "/chronograf/v1",
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/sources/{id}/kapacitors": {
      "get": {
        "tags": ["sources", "kapacitors"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "description": "ID of the source",
            "required": true
          }
        ],
        "summary": "Retrieve list of configured kapacitors",
        "responses": {
          "200": {
            "description": "An array of kapacitors",
            "schema": {
              "$ref": "#/definitions/Kapacitors"
            }
          },
          "default": {
            "description": "Unexpected internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "post": {
        "tags": ["sources", "kapacitors"],
        "summary": "Create new kapacitor backend",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "description": "ID of the source",
            "required": true
          },
          {
            "name": "kapacitor",
            "in": "body",
            "description": "Configuration options for kapacitor",
            "schema": {
              "$ref": "#/definitions/Kapacitor"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Kapacitor source successfully created",
            "headers": {
              "Location": {
                "type": "string",
                "format": "url",
                "description": "Location of the newly created kapacitor resource."
              }
            },
            "schema": {
              "$ref": "#/definitions/Kapacitor"
            }
          },
          "default": {
            "description": "A processing or an unexpected error.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/sources/{id}/kapacitors/{kapa_id}": {
      "get": {
        "tags": ["sources", "kapacitors"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "description": "ID of the source",
            "required": true
          },
          {
            "name": "kapa_id",
            "in": "path",
            "type": "string",
            "description": "ID of the kapacitor",
            "required": true
          }
        ],
        "summary": "Configured kapacitors",
        "description": "Retrieve information on a single kapacitor instance",
        "responses": {
          "200": {
            "description": "Kapacitor connection information",
            "schema": {
              "$ref": "#/definitions/Kapacitor"
            }
          },
          "404": {
            "description": "Unknown data source or kapacitor id",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Unexpected internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "patch": {
        "tags": ["sources", "kapacitors"],
        "summary": "Update kapacitor configuration",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "description": "ID of the source",
            "required": true
          },
          {
            "name": "kapa_id",
            "in": "path",
            "type": "string",
            "description": "ID of a kapacitor backend",
            "required": true
          },
          {
            "name": "config",
            "in": "body",
            "description": "kapacitor configuration",
            "schema": {
              "$ref": "#/definitions/Kapacitor"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Kapacitor's configuration was changed",
            "schema": {
              "$ref": "#/definitions/Kapacitor"
            }
          },
          "404": {
            "description": "Happens when trying to access a non-existent data source or kapacitor.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "A processing or an unexpected error.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "delete": {
        "tags": ["sources", "kapacitors"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "description": "ID of the source",
            "required": true
          },
          {
            "name": "kapa_id",
            "in": "path",
            "type": "string",
            "description": "ID of the kapacitor",
            "required": true
          }
        ],
        "summary": "Remove Kapacitor backend",
        "description": "This specific kapacitor will be removed. All associated rule resources will also be removed from the store.",
        "responses": {
          "204": {
            "description": "kapacitor has been removed."
          },
          "404": {
            "description": "Unknown Data source or Kapacitor id",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Unexpected internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Kapacitors": {
      "type": "object",
      "required": ["kapacitors"],
      "properties": {
        "kapacitors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Kapacitor"
          }
        }
      }
    },
    "Kapacitor": {
      "type": "object",
      "required": ["name", "url"],
      "example": {
        "id": "4",
        "name": "kapa",
        "url": "http://localhost:9092",
        "active": false,
        "insecureSkipVerify": false,
        "links": {
          "proxy": "/chronograf/v1/sources/4/kapacitors/4/proxy",
          "self": "/chronograf/v1/sources/4/kapacitors/4",
          "rules": "/chronograf/v1/sources/4/kapacitors/4/rules"
        }
      },
      "properties": {
        "id": {
          "type": "string",
          "description": "Unique identifier representing a kapacitor instance.",
          "readOnly": true
        },
        "name": {
          "type": "string",
          "description": "User facing name of kapacitor instance."
        },
        "username": {
          "type": "string",
          "description": "Username for authentication to kapacitor."
        },
        "password": {
          "type": "string",
          "description": "Password is in cleartext."
        },
        "url": {
          "type": "string",
          "format": "url",
          "description": "URL for the kapacitor backend (e.g. http://localhost:9092)"
        },
        "insecureSkipVerify": {
          "type": "boolean",
          "description": "True means any certificate presented by the kapacitor is accepted.  Typically used for self-signed certs. Probably should only be used for testing."
        },
        "active": {
          "type": "boolean",
          "description": "Indicates whether the kapacitor is the current kapacitor being used for a source"
        },
        "links": {
          "type": "object",
          "properties": {
            "self": {
              "type": "string",
              "description": "Self link mapping to this resource",
              "format": "url"
            },
            "proxy": {
              "type": "string",
              "description": "URL location of proxy endpoint for this kapacitor",
              "format": "url"
            },
            "rules": {
              "type": "string",
              "description": "URL location of rules endpoint for this kapacitor",
              "format": "url"
            }
          }
        }
      }
    },
...
    "Error": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
  }
}